ref-manual/lepton: update 'lepton-cli shell' page
lepton-eda PR #730:
"Fix a link in the reference manual".

TODO:
Relative links from the Lepton manual to the
Scheme manual have ../lepton-scheme.html/ prefix.
This will not work on the web site.
Most probably, we should move both so that
the directory hier be like the local one, e.g.:

/ref-manual/
|
|-- lepton-manual.html/
|-- lepton-scheme.html/